integration_tests/test-requirements.txt
git+https://github.com/wazo-pbx/xivo-test-helpers.git
git+https://github.com/wazo-pbx/xivo-provd-client.git
docker-compose
nose
pyhamcrest
psycopg2  # sqlalchemy engine
pyvirtualdisplay
requests
selenium
sqlalchemy
